pyproject.toml
[project]
name = "resume-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "pdfplumber>=0.11.9",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.24",
    "torch>=2.11.0",
    "transformers>=5.5.1",
    "uvicorn>=0.44.0",
]
